Changed input bg's from solid gray to faded white
When changing the page background color, I noticed that the search box in the sidebar widget was staying the gray #f7f7f7 color. With black or white backgrounds it was ok, but when you started using yellow or other color variations, the gray really clashed.

I believe this fixes the problem and allows the input fields to blend better with the theme, no matter what color the user selects.